requirements.txt
streamlit
pytesseract
opencv-python
openai
pillow
numpy
pdf2image
pdfplumber
requests
python-dotenv
